tests added; improvements

+++ backend/apps/django/authentication/otp/models/ip_geolocation.py
@@ -19,8 +19,8 @@ class IPGeolocation(IdMixin):
         verbose_name=_("Ip-Adresse"),
     )
     
-    longitude = models.PositiveSmallIntegerField()
-    latitude = models.PositiveSmallIntegerField()
+    longitude = models.FloatField()
+    latitude = models.FloatField()
     
     city = models.CharField(
         verbose_name=_("Stadt"),
